+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 2.0.2 Release Candidate
 
+### Bug fixes
+
+- Make provision user and group ids not rely on hardcoded values [#13192](https://github.com/opencrvs/opencrvs-core/issues/13192)
+
 ## 2.0.1 Release
 
 ### New features